% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_gpower.R
\name{auto_gpower}
\alias{auto_gpower}
\title{Find GPower rho for proportion of sparsity}
\usage{
auto_gpower(
  data,
  k,
  prop_sparse,
  accuracy = 2,
  reg = c("l0", "l1"),
  center = c(TRUE, FALSE),
  block = c(TRUE, FALSE),
  mu = 1,
  iter_max = 1000,
  epsilon = 1e-04
)
}
\arguments{
\item{data}{Input matrix of size (p x n) with p < n.}

\item{k}{Number of components, 0 < k < p.}

\item{prop_sparse}{The percentage of the total values of the weights matrix
which is equal to zero.}

\item{accuracy}{The amount of digits to which to round prop_sparse and the
sparsity of the weights.}

\item{reg}{regularisation type to use in the optimization. Either 'l0' or 'l1'.
The default is 'l1' since it performed best in experiments.}

\item{center}{Centers the data. Either TRUE or FALSE. The default is TRUE.}

\item{block}{Optimization method. If FALSE, the components are calculated
individually. If TRUE, all components are calculated at the same time.
The default is FALSE.}

\item{mu}{Mean to be applied to each component in the block. Either a vector
of float of size k or a float which will be repeated k times. Only used if
block is TRUE. The default is 1.}

\item{iter_max}{Maximum iterations when adjusting components with gradient
descent. The default is 1000.}

\item{epsilon}{Epsilon of the gradient descent stopping function.
The default is 1e-4.}
}
\description{
Using binary search find a value of rho for which the weights matrix of
\code{\link{gpower}} has a proportion of sparsity close to
prop_sparse. It forwards the other settings to the \code{\link{gpower}}
function.
}
\examples{
set.seed(360)
p <- 20
n <- 50
k <- 5
data <- matrix(stats::rnorm(p * n), nrow = p, ncol = n)
prop_sparse <- 0.1
mu <- 1

auto_gpower(data, k, prop_sparse, reg = 'l1', center = TRUE, block = FALSE)

}
