% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{gpower_comp_var_plot}
\alias{gpower_comp_var_plot}
\title{Plots explained variance of each component}
\usage{
gpower_comp_var_plot(
  data,
  k,
  intervals,
  penalty = "l1",
  center = TRUE,
  block = FALSE,
  mu = 1
)
}
\arguments{
\item{data}{Input matrix of size (p x n) with p < n.}

\item{k}{Number of components, 0 < k < p.}

\item{intervals}{The amount of intervals in the range of values of rho for
which gpower is run. For l1, the range of values is 0-1, for l0, the range
is 0-0.33. The block algorithms tend to stop at 0.4 and will only cover the
range where gpower is able to run.}

\item{penalty}{Penalty type to use in the optimization. Either 'l0' or 'l1'.
The default is 'l1' since it performed best in experiments.}

\item{center}{Centers the data. Either TRUE or FALSE. The default is TRUE.}

\item{block}{Optimization method. If FALSE, the components are calculated
individually. If TRUE, all components are calculated at the same time.
The default is FALSE.}

\item{mu}{Mean to be applied to each component in the block. Either a vector
of float of size k or a float which will be repeated k times. Only used if
block is TRUE. The default is 1.}
}
\description{
This plot shows what happens to the variance explained by each component if
rho is increased. The individual explained variances are neither
non-decreasing nor non-increasing. Their sum does trend downwards, but it is
also not non-increasing just like the proportion of explained variance.
}
\examples{
set.seed(360)
p <- 20
n <- 50
k <- 5
data <- scale(matrix(stats::rnorm(p * n), nrow = p, ncol = n), scale = FALSE)

gpower_comp_var_plot(
  data = data,
  k = k,
  intervals = 40,
  penalty = 'l1',
  center = TRUE
)
}
