% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpower.R
\name{print.gpower}
\alias{print.gpower}
\title{Prints the proportion of explained variance and the weights of the gpower
object.}
\usage{
\method{print}{gpower}(x, print_zero_rows = TRUE, ...)
}
\arguments{
\item{x}{gpower object}

\item{print_zero_rows}{If FALSE, then only the rows which contain at least 1
non-zero value will be printed.}

\item{...}{print() parameters}
}
\description{
Prints the proportion of explained variance and the weights of the gpower
object.
}
