% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpower.R
\name{summary.gpower}
\alias{summary.gpower}
\title{Prints the proportion of explained variance and the proportion of sparseness.}
\usage{
\method{summary}{gpower}(object, ...)
}
\arguments{
\item{object}{gpower object}

\item{...}{summary() parameters}
}
\description{
Prints the proportion of explained variance and the proportion of sparseness.
}
