% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{gpower_component_heatmap}
\alias{gpower_component_heatmap}
\title{Plots value of each column of the data for each component}
\usage{
gpower_component_heatmap(
  data,
  k,
  rho,
  penalty = "l1",
  center = TRUE,
  block = FALSE,
  mu = 1,
  variable_highlight = NA,
  cluster_variables = FALSE,
  show_variable_names = TRUE
)
}
\arguments{
\item{data}{Input matrix of size (p x n) with p < n.}

\item{k}{Number of components, 0 < k < p.}

\item{rho}{Relative sparsity weight factor of the optimization. Either a
vector of floats of size k or float which will be repeated k times. 0 < rho
< 1.}

\item{penalty}{Penalty type to use in the optimization. Either 'l0' or 'l1'.
The default is 'l1' since it performed best in experiments.}

\item{center}{Centers the data. Either TRUE or FALSE. The default is TRUE.}

\item{block}{Optimization method. If FALSE, the components are calculated
individually. If TRUE, all components are calculated at the same time.
The default is FALSE.}

\item{mu}{Mean to be applied to each component in the block. Either a vector
of float of size k or a float which will be repeated k times. Only used if
block is TRUE. The default is 1.}

\item{variable_highlight}{Add a color coding to the variables using a matrix with of size n x 1 where the row names are the same as the column names of the data matrix.}

\item{cluster_variables}{Cluster the variables using hierarchical clustering.}

\item{show_variable_names}{Show the names of the variables on the right side of the graph.}
}
\description{
This plot shows the values of the components, connected to the columns of the
original data. The heatmap is made using the pheatmap function from the pheatmap package.
}
\examples{
set.seed(360)
p <- 20
n <- 50
k <- 5
data <- scale(matrix(stats::rnorm(p * n), nrow = p, ncol = n), scale = FALSE)

gpower_component_heatmap(
  data = data,
  k = 5,
  rho = 0.1,
  penalty = 'l1',
  center = TRUE,
  block = FALSE,
  mu = 1,
  variable_highlight = NA,
  cluster_variables = FALSE,
  show_variable_names = TRUE
)
}
